Finished moving things around in SiteTable, not sure how I feel about it
